add text output stream log writer
